+++ docs/CHANGELOG.md
@@ -15,6 +15,14 @@ Changes are grouped as follows:
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## TBD
+### Added
+- Support for information rules and instance plotting in NeatSession
+
+### Improved
+- Better plotting of rules for dms and information rules in NeatSession (accounts for `subClassOf` and `implements`)
+
+
 ## [0.93.0] - 28-10-2024
 ### Improved
 - IODD extractor to also extract XML elements that should map to time series data type